# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /members/{id}/bills:
    get:
      operationId: listMemberBills
      tags: [members, bills]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
  /members/{id}/attendance:
    get:
      operationId: listMemberAttendance
      tags: [members, attendance]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /members/{id}/body-analysis:
    get:
      operationId: listMemberBodyAnalysis
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BodyAnalysis"
    post:
      operationId: createMemberBodyAnalysis
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBodyAnalysisInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyAnalysis"
  /members/{id}/notes:
    get:
      operationId: listMemberNotes
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberNote"
    post:
      operationId: createMemberNote
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberNote"
  /members/{id}/documents:
    get:
      operationId: listMemberDocuments
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberDocument"
  /members/{id}/workout-plans:
    get:
      operationId: listMemberWorkoutPlans
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
  /members/{id}/nutrition-plans:
    get:
      operationId: listMemberNutritionPlans
      tags: [members]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NutritionPlan"
...
  /pos/sales:
    get:
      operationId: listPosSales
      tags: [pos]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PosSale"
    post:
      operationId: createPosSale
      tags: [pos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePosSaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosSale"
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, name, phone, status, plan, joinDate, expiryDate, gender, avatarUrl, balanceDue]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        plan: { type: string }
        joinDate: { type: string, format: date-time }
        expiryDate: { type: string, format: date-time }
        gender: { type: string }
        bloodGroup: { type: ["string", "null"] }
        dob: { type: ["string", "null"], format: date-time }
        address: { type: ["string", "null"] }
        avatarUrl: { type: string }
        balanceDue: { type: integer }
        medicalHistory: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
    CreateMemberInput:
      type: object
      required: [name, phone, gender, plan]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        gender: { type: string }
        bloodGroup: { type: ["string", "null"] }
        dob: { type: ["string", "null"], format: date-time }
        address: { type: ["string", "null"] }
        plan: { type: string }
        medicalHistory: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
    UpdateMemberInput:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        gender: { type: string }
        bloodGroup: { type: ["string", "null"] }
        dob: { type: ["string", "null"], format: date-time }
        address: { type: ["string", "null"] }
        plan: { type: string }
        status: { type: string }
        expiryDate: { type: string, format: date-time }
        medicalHistory: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
...
    AttendanceRecord:
      type: object
      required: [id, memberId, memberName, checkIn, source]
      properties:
        id: { type: string }
        memberId: { type: string }
        memberName: { type: string }
        memberAvatar: { type: ["string", "null"] }
        checkIn: { type: string, format: date-time }
        checkOut: { type: ["string", "null"], format: date-time }
        source: { type: string }
...
    Bill:
      type: object
      required: [id, memberId, memberName, type, amount, paid, balance, status, createdAt]
      properties:
        id: { type: string }
        memberId: { type: string }
        memberName: { type: string }
        type: { type: string }
        items:
          type: array
          items:
            type: object
            required: [name, amount]
            properties:
              name: { type: string }
              amount: { type: integer }
        amount: { type: integer }
        paid: { type: integer }
        balance: { type: integer }
        status: { type: string }
        paymentMode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    PosSale:
      type: object
      required: [id, items, subtotal, tax, total, paymentMode, createdAt]
      properties:
        id: { type: string }
        memberId: { type: ["string", "null"] }
        memberName: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [productId, name, quantity, price]
            properties:
              productId: { type: string }
              name: { type: string }
              quantity: { type: integer }
              price: { type: integer }
        subtotal: { type: integer }
        tax: { type: integer }
        total: { type: integer }
        paymentMode: { type: string }
        createdAt: { type: string, format: date-time }
    CreatePosSaleInput:
      type: object
      required: [items, paymentMode]
      properties:
        memberId: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string }
              quantity: { type: integer }
        paymentMode: { type: string }
...
    BodyAnalysis:
      type: object
      required: [id, memberId, recordedAt, weightKg, bodyFatPct]
      properties:
        id: { type: string }
        memberId: { type: string }
        recordedAt: { type: string, format: date-time }
        weightKg: { type: number }
        bodyFatPct: { type: number }
        muscleKg: { type: ["number", "null"] }
        bmi: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    CreateBodyAnalysisInput:
      type: object
      required: [weightKg, bodyFatPct]
      properties:
        weightKg: { type: number }
        bodyFatPct: { type: number }
        muscleKg: { type: ["number", "null"] }
        bmi: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
    MemberNote:
      type: object
      required: [id, memberId, body, author, createdAt]
      properties:
        id: { type: string }
        memberId: { type: string }
        body: { type: string }
        author: { type: string }
        createdAt: { type: string, format: date-time }
    CreateMemberNoteInput:
      type: object
      required: [body]
      properties:
        body: { type: string }
        author: { type: ["string", "null"] }
    MemberDocument:
      type: object
      required: [id, memberId, name, type, status, uploadedBy, uploadedAt]
      properties:
        id: { type: string }
        memberId: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        uploadedBy: { type: string }
        uploadedAt: { type: string, format: date-time }
        note: { type: ["string", "null"] }
    WorkoutPlan:
      type: object
      required: [id, memberId, name, weeks, days, trainerNote, active]
      properties:
        id: { type: string }
        memberId: { type: string }
        name: { type: string }
        weeks: { type: string }
        active: { type: boolean }
        trainerNote: { type: string }
        days:
          type: array
          items:
            type: object
            required: [day, exercises]
            properties:
              day: { type: string }
              exercises:
                type: array
                items:
                  type: object
                  required: [name, setsReps]
                  properties:
                    name: { type: string }
                    setsReps: { type: string }
    NutritionPlan:
      type: object
      required: [id, memberId, name, calories, active, macros, meals]
      properties:
        id: { type: string }
        memberId: { type: string }
        name: { type: string }
        calories: { type: integer }
        active: { type: boolean }
        macros:
          type: object
          required: [protein, carbs, fats]
          properties:
            protein: { type: string }
            carbs: { type: string }
            fats: { type: string }
        meals:
          type: array
          items:
            type: object
            required: [name, time, kcal, description]
            properties:
              name: { type: string }
              time: { type: string }
              kcal: { type: integer }
              description: { type: string }
...