# lib/api-zod/src/generated/types/bodyAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BodyAnalysis {
  id: string;
  memberId: string;
  recordedAt: Date;
  weightKg: number;
  bodyFatPct: number;
  /** @nullable */
  muscleKg?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  notes?: string | null;
}